Refactor license summary
